% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_checkyn.R
\name{ba_checkyn}
\alias{ba_checkyn}
\title{Get an arbitrary answer to a question}
\usage{
ba_checkyn(question)
}
\arguments{
\item{question}{the question for which you want an answer}
}
\description{
Get an arbitrary answer to a question
}
\examples{
ba_checkyn("how are you feeling?")
}
\keyword{question}

