% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ba_indexFromPlatePos.R
\name{ba_indexFromPlatePos}
\alias{ba_indexFromPlatePos}
\title{# e.g. L9 in '20131016_Oxford_Samples.L9'Get an arbitrary answer to a question}
\usage{
ba_indexFromPlatePos(data, posString, verbose = F)
}
\arguments{
\item{data}{the MALDIquant object holding the bruker data}

\item{posString}{the spot postition, e.g. 'C9'}

\item{verbose}{verbose output if True}
}
\description{
# e.g. L9 in '20131016_Oxford_Samples.L9'Get an arbitrary answer to a question
}
\examples{
ba_indexFromPlatePos("/home/anon/brukerdata/20151202",'C9')
}
\keyword{MALDIquant}

