% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ba_loadBruker.R
\name{ba_loadBruker}
\alias{ba_loadBruker}
\title{load a bruker directory structure as a dataset}
\usage{
ba_loadBruker(xml)
}
\arguments{
\item{xml}{the path to the xml file on the computer you are working at}
}
\description{
load a bruker directory structure as a dataset
}
\examples{
ba_loadBruker("/home/anon/brukerdata/20151202")
}
\keyword{bruker}

