% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ba_logspot.R
\name{ba_logspot}
\alias{ba_logspot}
\title{Create a structure that holds the q2e parameters}
\usage{
ba_logspot(spot)
}
\arguments{
\item{spot}{the spot address (e.g. "A12")}
}
\description{
Create a structure that holds the q2e parameters
}
\examples{
ba_logspot("A12")
}
\keyword{platemap}

