% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ba_ynq.R
\name{ba_ynq}
\alias{ba_ynq}
\title{Get an arbitrary answer to a yes/no question}
\usage{
ba_ynq(question)
}
\arguments{
\item{question}{the question for which you want a yes/no answer}
}
\description{
Get an arbitrary answer to a yes/no question
}
\examples{
ba_ynq("are you well?")
}
\keyword{question}

