% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioarch_exportDataSetTab.R
\name{bioarch_exportDataSetTab}
\alias{bioarch_exportDataSetTab}
\title{Export Bruker data to files, with lot Number as part of the filename
this function will create files either with the file extension .txt, or using the fullname in the Bruker format.
NB: make sure you have created a subdirectory called 'csv' from your working directory!}
\usage{
bioarch_exportDataSetTab(data, txt = TRUE)
}
\arguments{
\item{path}{the path to the Bruker data set}
}
\description{
Export Bruker data to files, with lot Number as part of the filename
this function will create files either with the file extension .txt, or using the fullname in the Bruker format.
NB: make sure you have created a subdirectory called 'csv' from your working directory!
}
\examples{
1: to export directly from a Bruker directory:

  > bioarch_exportDataSetTab(bioarch_loadBruker("path/to/bruker/data"))

2: to export from a Bruker data set that has been loaded into an
object called 'bdata':

  > bioarch_exportDataSetTab(bdata)

3: to export from a Bruker object called 'bdata' to files named
after the Bruker 'fullname':

  > bioarch_exportDataSetTab(bdata,txt=T)
}
\keyword{Bruker}

