% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioarch_loadBrukerXML.R
\name{bioarch_loadBrukerXML}
\alias{bioarch_loadBrukerXML}
\title{load Bruker dataset via xml}
\usage{
bioarch_loadBrukerXML(xml)
}
\description{
load Bruker dataset via xml
}
\examples{
data <- bioarch_loadBrukerXML("Bruker.xml")
}
\keyword{MALDIquant}
\keyword{platemap}

