% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportBruker.R
\name{exportDataSetTab}
\alias{exportDataSetTab}
\title{Export Bruker data to files, with lot Number as part of the filename}
\usage{
exportDataSetTab(data, txt = TRUE)
}
\arguments{
\item{path}{the path to the Bruker data set}
}
\description{
Export Bruker data to files, with lot Number as part of the filename
}
\examples{
exportDataSetTab("/home/anon/Bruker/20151105sample")
}
\keyword{Bruker}

