% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_lotspot.R
\name{exportLotData}
\alias{exportLotData}
\title{Export Bruker data to files, with lot Number as part of the filename}
\usage{
exportLotData(path, lotno = 0, txt = TRUE)
}
\arguments{
\item{path}{the path to the Bruker data set}

\item{lotno}{the lot number of the data set (checked in the function)}
}
\description{
Export Bruker data to files, with lot Number as part of the filename
}
\examples{
exportLotData("/home/anon/Bruker/20151105sample",123456)
}
\keyword{Bruker}

