% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ba_lotspot.R
\name{exportLotDataSet}
\alias{exportLotDataSet}
\title{Get an arbitrary answer to a question}
\usage{
exportLotDataSet(data, txt = TRUE)
}
\arguments{
\item{data}{the path to the Bruker data set}
}
\description{
Get an arbitrary answer to a question
}
\examples{
ba_checkyn("how are you feeling?")
}
\keyword{question}

