% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportBruker.R
\name{loadBrukerXML}
\alias{loadBrukerXML}
\title{load Bruker dataset via xml}
\usage{
loadBrukerXML(xml)
}
\description{
load Bruker dataset via xml
}
\examples{
zloadBruker("Bruker.xml")
}
\keyword{MALDIquant}
\keyword{platemap}

