% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flamePlot_3.R
\name{plotclass_v2}
\alias{plotclass_v2}
\title{Plot a classification of a Bruker data set}
\usage{
plotclass_v2(fn, onepdf = F, refdata, testdata, threshold, uselag = TRUE,
  thislb, thisub, lbl, ubl, classnames)
}
\arguments{
\item{fn}{the pdf file name}

\item{onepdf}{whether to print the analysis into a single file or not}

\item{refdata}{the reference data that the classification is based on}

\item{testdata}{the data that is being classified}

\item{threshold}{classification threshold}

\item{uselag}{whether to calculate the lag}

\item{thislb}{lower bound 1}

\item{thisub}{upper bound 1}

\item{lbl}{lower bound limit}

\item{ubl}{upper bound limit}

\item{classnames}{the names of the classes}
}
\description{
Plot a classification of a Bruker data set
}
\examples{
ba_checkyn("how are you feeling?")
}
\keyword{question}

