% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportBruker.R
\name{printUsage}
\alias{printUsage}
\title{print the usage of exportBruker}
\usage{
printUsage()
}
\description{
print the usage of exportBruker
}
\examples{
printUsage()
}
\keyword{MALDIquant}
\keyword{platemap}

