% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ba_indexFromPlatePos.R
\name{ba_indexFromPlatePos}
\alias{ba_indexFromPlatePos}
\title{Get the index of a platemap location in a MALDIquant data object}
\usage{
ba_indexFromPlatePos(data, posString, verbose = F)
}
\arguments{
\item{data}{the MALDIquant object holding the bruker data}

\item{posString}{the spot postition, e.g. 'C9'}

\item{verbose}{verbose output if True}
}
\description{
Get the index of a platemap location in a MALDIquant data object
}
\examples{
ba_indexFromPlatePos("/home/anon/brukerdata/20151202",'C9')
}
\keyword{MALDIquant}

