% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_logspot.R
\name{ba_logspot}
\alias{ba_logspot}
\title{DEPRECIATED: Log a new spot position}
\usage{
ba_logspot(spot)
}
\arguments{
\item{spot}{the spot address (e.g. "A12")}
}
\description{
DEPRECIATED: Log a new spot position
}
\examples{
ba_logspot("A12")
}
\keyword{platemap}

