% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportBruker.R
\name{printUsage}
\alias{printUsage}
\title{DEPRECIATED print the usage of exportBruker}
\usage{
printUsage()
}
\description{
DEPRECIATED print the usage of exportBruker
}
\examples{
printUsage()
}
\keyword{Bruker}
\keyword{MALDIquant}
\keyword{platemap}

